feat: Update favicon to new design
This commit updates the favicon to a new design. The previous favicon file, vite.svg, has been replaced with a new file, favicon.svg, which contains the updated design. The changes were made in the client/public directory. This update improves the visual representation of the application and aligns with the new branding guidelines.