feat: Add route for generating 204 response
This commit adds a new route '/generate_204' to the server. The route is a GET request handler that returns an empty response with a status code of NO_CONTENT. This route is used to check if the user is connected to the internet and if the API endpoint is live. The addition of this route enhances the functionality of the application.